+++ apps/docs/src/guide/contracts/logs.md
@@ -4,10 +4,10 @@ When you log a value within a contract method, it generates a log entry that is 
 
 Consider the following example contract:
 
-<<< @/../../docs-snippets/test/fixtures/forc-projects/log-values/src/main.sw#log-1{rust:line-numbers}
+<<< @/../../docs-snippets2/sway/log-values/src/main.sw#log-1{rust:line-numbers}
 
 To access the logged values in TypeScript, use the `logs` property in the response of a contract call. The logs data will be stored in an `Array<any>`:
 
-<<< @/../../docs-snippets/src/guide/contracts/logs.test.ts#log-2{ts:line-numbers}
+<<< @/../../docs-snippets2/src/contracts/logs.ts#full{ts:line-numbers}
 
 This approach allows you to work seamlessly with logged values in your contract, making it easier to understand and debug the contract's behavior.
